+++ test/unit/Parameter/Injector/FlatInjectorTest.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Test\Unit\Krizalys\Onedrive\Parameter\Injector;
 
 use Krizalys\Onedrive\Parameter\Injector\FlatInjector;