+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Change log
 
 - Support for PHP 5.6, PHP 7.0, PHP 7.1, and PHP 7.2.
 - V1 API.
+- V2 API.
 
 [2.7.0] - 2021-02-12
 --------------------